Updates for brightness toggling via Hold
I've made a small array for 3 brightness settings to be toggled by holding the button.  I've preserved the original "auto mode" via 3 button presses in 2 seconds, and added an "off" mode when pressed 5 times in 2 seconds.  Personally I like the GPIO 0 as a button pin because it makes these projects insanely easy and cheap, but to each his own.